backend/requirements.txt
# Core Django
django==5.1.*
djangorestframework==3.15.*
django-cors-headers==4.6.*

# Multi-tenancy
django-tenants==3.7.*

# Authentication
djangorestframework-simplejwt==5.3.*
django-allauth==65.*

# Database
psycopg2-binary==2.9.*
dj-database-url==2.2.*
django-redis==5.4.*

# Task queue
celery==5.4.*
redis==5.0.*
django-celery-beat==2.7.*

# API tooling
django-filter==24.*
drf-spectacular==0.27.*

# Payments
razorpay==1.4.*

# PDF generation
WeasyPrint==62.*

# Excel export
openpyxl==3.1.*

# Image processing
Pillow==11.*
qrcode[pil]==8.*

# Utilities
python-decouple==3.8
structlog==24.*
django-ratelimit==4.1.*

# Development only
django-debug-toolbar==4.4.*
